backend/requirements.txt
flask
Flask[async]
flask_cors
python-dotenv
requests_oauthlib
axios
aiohttp